Core/Players: Fixed player model weirdness with teleports/shapeshifts/transforms
Closes #18048

+++ src/server/game/Entities/Creature/Creature.h
@@ -548,7 +548,7 @@ struct CreatureModelInfo
 };
 
 // Benchmarked: Faster than std::map (insert/find)
-typedef std::unordered_map<uint16, CreatureModelInfo> CreatureModelContainer;
+typedef std::unordered_map<uint32, CreatureModelInfo> CreatureModelContainer;
 
 enum InhabitTypeValues
 {
